feat(lambda): move Input exception into appropriate directory #59
